refactor(Instantiater.kt): 전체적인 리팩터링
